+++ docs/docs/changelog.md
@@ -3,7 +3,9 @@ pdf:
   filename: Changelog
   title: Changelog & License for MkDocs PDF Generate Plugin
   type: Changelog
-  revision: 0.2.0   
+  revision: 0.2.1   
+  toc_txt: true
+  csv_name: Plugin Changelog
 ---
 
 # Changelog & License
@@ -34,9 +36,21 @@ SOFTWARE.
 
 # Changelog
 
+0.2.1
+-----
+Released: **22-03-2023**
+
+* Updated documentation and made minor changes to plugin code.
+* Added the `csv_name` local option.
+* Improvement: Added feature to generate a CSV file for all documents with the `toc_txt` local option set to `true`.
+* Improvement: Added changes that appends the revision number of a document to the filename.
+* Bugfix: Added default URL to use if `site_url` is not defined in mkdocs.yml. Fix for issue [#15](https://github.com/iSOLveIT/mkdocs-pdf-generate/issues/15).
+* Improvement: Added the `toc_txt` local option that allows users to build TXT file containing the Table of Contents
+* Minor improvements: Rewrote logging messages.
+
 0.2.0
 -----
-Released: 14-02-2023
+Released: **14-02-2023**
 
 * Package setup: Changed package from using setuptools to pyproject.toml